==> src/1-DAQ.ipynb <==
In order to create the annual estimate of wildfire smoke in Salina, Kansas, I am going to read the USGS_Wildland_Fire_Combined_Dataset.json which has been downloaded from the [USGS Website from the Combined wildland fire datasets for the United States and certain territories, 1800s-Present (combined wildland fire polygons)](https://www.sciencebase.gov/catalog/item/61aa537dd34eb622f699df81) dataset. 

In order to explore the dataset, I will be reading the JSON data with the help of [this](https://drive.google.com/file/d/1TwCkvdaw0MxJzW7NSDg6XxYQ0dvaS44I/view) solution provided by Professor McDonald. I will also be using code from the [Wildfire Proximity Computational Example notebook provided](https://colab.research.google.com/drive/1qNI6hji8CvDeBsnLDAhJXvaqf2gcg8UV#scrollTo=DwLfwdk9rXcd). The notebook is licensed for use in DATA 512, a course in the UW MS Data Science degree program. This code is provided under the Creative Commons CC-BY license. Revision 1.0 - August 13, 2023

# Section 1: Importing Libraries and Initial Setup

I'm starting by bringing in some essential tools to help with our data work. We've got Pandas for handling data, geospatial libraries for calculating distances, and a custom module we obtained from our course material to deal with wildfire data.

import pandas as pd
import json, os, time
from geopy.distance import geodesic
import sys, json
from wildfire.Reader import Reader as WFReader
from pyproj import Transformer, Geod
import geojson

- You might need to install pyproj and geojson, you can do so via pip.
- The 'wildfire' module has been taken from the course drive and can be found [here](https://drive.google.com/file/d/1TwCkvdaw0MxJzW7NSDg6XxYQ0dvaS44I/view).
- This module includes a Reader object for reading GeoJSON files linked to the wildfire dataset.
- It also provides a sample datafile called Wildfire_short_sample.json that contains California wildfires data.

# Section 2: Coordinate Conversion Function

We firstly take the geometry of a fire feature, extract the largest ring (i.e., the largest boundary of the fire) and convert all of the points in that ring from the ESRI:102008 coordinate system to EPSG:4326 coordinates. This is going to be useful for our geospatial work.

def convert_ring_to_epsg4326(ring_data=None):
    converted_ring = list()
    
    to_epsg4326 = Transformer.from_crs("ESRI:102008","EPSG:4326")
    for coord in ring_data:
        lat,lon = to_epsg4326.transform(coord[0],coord[1])
        new_coord = lat,lon
        converted_ring.append(new_coord)
    return converted_ring

# Section 3: Finding Shortest Distance to Fire Perimeter

Our next is step is finding shortest distance between Salina and each wildfire. The following code finds the point on the perimiter with the shortest distance to the city (place) and returns the distance as well as the lat,lon of the perimeter point.   

def shortest_distance_from_place_to_fire_perimeter(place=None,ring_data=None):
    ring = convert_ring_to_epsg4326(ring_data)    
    geodcalc = Geod(ellps='WGS84')
    closest_point = list()
    for point in ring:
        d = geodcalc.inv(place[1],place[0],point[1],point[0])
        distance_in_miles = d[2]*0.00062137
        if not closest_point:
            closest_point.append(distance_in_miles)
            closest_point.append(point)
        elif closest_point and closest_point[0]>distance_in_miles:
            closest_point = list()
            closest_point.append(distance_in_miles)
            closest_point.append(point)
    return closest_point

I am going to read the GeoJSON data, USGS_Wildland_Fire_Combined_Dataset.json, which has been downloaded from the [USGS Website from the Combined wildland fire datasets for the United States and certain territories, 1800s-Present (combined wildland fire polygons)](https://www.sciencebase.gov/catalog/item/61aa537dd34eb622f699df81) dataset. 

For reference, SAMPLE_DATA_FILENAME is the path to USGS_Wildland_Fire_Combined_Dataset.json 


# Section 4: Loading GeoJSON Data

SAMPLE_DATA_FILENAME = '/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/USGS_Wildland_Fire_Combined_Dataset.json'

Pulling metadata to see what we're working with!

print(f"Attempting to open '{SAMPLE_DATA_FILENAME}' with wildfire.Reader() object")
wfreader = WFReader(SAMPLE_DATA_FILENAME)
print()

header_dict = wfreader.header()
header_keys = list(header_dict.keys())
#print("The header has the following keys:")
#print(gj_keys)
print()
print("Header Dictionary")
print(json.dumps(header_dict,indent=4))

# Section 5: Loading and Processing Fire Features

Here, I'm looping through all the wildfire features in our dataset and storing them in a list. We're also keeping track of how many we've processed! 

MAX_FEATURE_LOAD = 100
feature_list = list()
feature_count = 0

wfreader.rewind()
feature = wfreader.next()
while feature:
    feature_list.append(feature)
    feature_count += 1
    if (feature_count % 100) == 0:
        print(f"Loaded {feature_count} features")
    '''if feature_count >= MAX_FEATURE_LOAD:
        break'''
    feature = wfreader.next()
print(f"Loaded a total of {feature_count} features")

print(f"Variable 'feature_list' contains {len(feature_list)} features")

# Section 6: Defining City Data

Specifying the coordinates for the city of Salina, Kansas

CITY_DATA = {
    'salina' :     {'city'   : 'salina', 
                    'latlon' : [38.8403, -97.6114]}}

# Section 7: Calculating Distances to Fire Perimeter

Calculating the distance from the closest edge of the fire to the town.

city_data = CITY_DATA["salina"]
fire_ids = []
shortest_dist_from_edge_list = []
features_processed = 0

There are some fires which aren't in ring shape. These are being processed in the try/except block to catch and deal with them.

for wf_feature in feature_list:
    try:
        ring_data = wf_feature['geometry']['rings'][0]
        distance = shortest_distance_from_place_to_fire_perimeter(city_data['latlon'], ring_data)
        fire_ids.append(wf_feature['attributes']['OBJECTID'])
        shortest_dist_from_edge_list.append(round(distance[0], 2))
    except KeyError:
        print("{0} fire has an irregular shape; skipping.".format(wf_feature['attributes']['OBJECTID']))

    features_processed += 1
    if features_processed % 1000 == 0:
        print("Processed {0} features".format(features_processed))
    if features_processed % 10000 == 0:
        dist_df = pd.DataFrame({'OBJECTID': fire_ids, 'shortest_dist': shortest_dist_from_edge_list})
        dist_df.to_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/distance.csv', index=False)

distance_df = pd.DataFrame({'OBJECTID': fire_ids, 'shortest_dist': shortest_dist_from_edge_list})
distance_df.to_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/distance.csv', index=False)

Reading the distances df

distance_df = pd.read_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/distance.csv')

We will now use the following conditions to further filter the data. Your smoke estimate should adhere to the following conditions:  
- The estimate only considers the last 60 years of wildland fires (1963-2023).
- The estimate only considers fires that are within 1250 miles of your assigned city.
- An annual fire season will run from May 1st through October 31st.


First we keep only fires which occurr within 1250 miles from Salina.

miles_df = distance_df.loc[distance_df['shortest_dist']< 1250]
print("{0} fires occured within 1250 miles of Salina".format(len(miles_df)))

miles_df.to_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/miles.csv', index = False)

Now we further filter this by keeping only fires which occurred in or after 1963. Since the data we have is avaliable till 2020 we don't need to specify that.

wf_file = open('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/USGS_Wildland_Fire_Combined_Dataset.json')
wf_dict = json.load(wf_file)

# Section 8: Filter and Extract Fire Data for years 1963 and after

We now filter and extract specific fire data entries with a 'Fire_Year' attribute greater than or equal to 1963 and stores relevant information in separate lists.

obj_id = []
fire_types = []
fire_years = []
acres = []
overlap_flags = []
fire_count = 0

for fire in wf_dict['features']:
    if fire['attributes']['Fire_Year'] >= 1963:
        obj_id.append(fire['attributes']['OBJECTID'])
        fire_types.append(fire['attributes']['Assigned_Fire_Type'])
        fire_years.append(fire['attributes']['Fire_Year'])
        acres.append(fire['attributes']['GIS_Acres'])
        overlap_flags.append(fire['attributes']['Overlap_Within_1_or_2_Flag'])
        fire_count += 1

        if fire_count % 1000 == 0:
            print("Processed {0} fires".format(fire_count))

feature_df = pd.DataFrame({'OBJECTID':obj_id,
                            'FireType':fire_types,
                            'FireYear':fire_years,
                            'GISAcres':acres,
                            'OverlapFlags' : overlap_flags})
feature_df.to_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/features.csv', index=False)

feature_df 

feature_df = pd.read_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/features.csv')
miles_df= pd.read_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/miles.csv')

Next we inner join the post-1963 fire attributes with their distances to calculate the smoke estimate. We will save this dataframe for later use in the data_processing script.

miles_df

feature_df

# Section 9: Data Checks

Lastly, we're checking and reporting some statistics about the fire data. We're looking at things like the most recent year, the earliest year, and the farthest fire from Salina. This helps us understand our data better.

filter_df = pd.merge(feature_df, miles_df, how = 'inner', left_on='OBJECTID', right_on='OBJECTID')
filter_df.to_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/filter.csv', index=False)

max_year = filter_df['FireYear'].max()
min_year = filter_df['FireYear'].min()
max_dist = filter_df['shortest_dist'].max()

if max_year > 2023:
    print("Fires that occured after 2023 are present.")
else:
    print("Maximum fire year: {0}.".format(max_year))

if min_year < 1963:
    print("Fires that occured before 1963 are present.")
else:
    print("Minimum fire year: {0}.".format(min_year))

if max_dist > 1250:
    print("Fires that occured more than 1250 miles away from Salina are present.")
    print("The farthest fire is {0} miles from Salina.".format(max_dist))
    print("The closest fire is {0} miles from Salina.".format(filter_df['shortest_dist'].min()))

num_fires = len(filter_df)
print("There are {0} fires after 1963 within 1250 miles of Salina.".format(num_fires))

filter_df
