==> src/2-Cleaning.ipynb <==
The contents of this notebook show how I've come up with my Smoke Estimate.

# Section 1: Importing Libraries and Initial Setup

import pandas as pd
import re

filter_df = pd.read_csv('/Users/aviva/Desktop/MSDS/quarter_4/Human Centered Data Science/Part 1/GeoJSON Exports/intermediate_data/filter.csv')

Trying to see the contents of my Overlap Flag column to see how I can leverage it.

print(filter_df['OverlapFlags'].unique())

filter_df.head()

The objective is now to estimate the impact of smoke from a wildfire on Salina.

We have two primary pieces of information: the number of acres burned by the wildfire `GISAcres` and the distance between the wildfire and the city `shortest_dist`.     

The first step is to acknowledge that several elements, including as plant type, fire intensity, wind direction, weather, and terrain, affect how much smoke is created and spread. This intricate problem is made simpler by introducing a proportional connection. It is expected that the amount of smoke that reaches the city is inversely related to the distance from the fire to the city and directly proportional to the size of the fire (measured in acres).   

`Smoke Factor`  
To account for the variation in smoke production per acre due to factors like vegetation type and fire intensity, a constant factor called Smoke_Factor is introduced. This factor provides a simplification of various elements that affect the amount of smoke produced per acre.  

This factor takes into account Acres Burned and Distance from Salina. More about this:  

`Acre Classification`  
We categorize wildfires into three size categories: "small fire," "medium fire," and "large fire." These categories are determined based on the number of acres burned by the wildfire. For example, a "small fire" could represent fires with fewer acres burned, while a "large fire" could represent massive fires.  

`Distance Classification`    
Wildfires are also categorized by their proximity to the city, resulting in classifications like "close fire," "intermediate fire," and "far fire."   

Based on these classification values, we then define a mapping of Smoke_Factor based on combinations of acre classification and dist classification. For instance:  

- A "small fire" that is "close to the city" may have a Smoke Factor of 0.1.  
- A "medium fire" that is "far from the city" may have a Smoke Factor of 0.3.  
- A "large fire" that is "close to the city" may have a Smoke Factor of 0.9.  

`Overlap Factor`     
We also introduce an Overlap_Factor, which considers the cumulative effects of previous fires and the extent to which they overlap with the current fire's location and time. This factor helps account for how multiple fires, over time and space, may contribute to the overall smoke production and dispersion. This factor takes into account both the time since the previous burn `Years_Since_Previous_Burn` and the extent of the overlap between the wildfires `Overlap_Percentage` which are extracted from the `OverlapFlags` column.    

**Finally, we construct the formula:**.      

`Smoke_Estimate = (Burned_Acres / Distance_from_City) * Smoke_Factor * (1 + Overlap_Factor)`   

The Overlap_Factor is computed using information from the 'OverlapFlags' column. If no overlap information is available, it is set to 0. If overlap information is present, the code extracts the years since the previous burn and the overlap percentage. The Overlap_Factor is then computed as    

`(Years_Since_Previous_Burn + 1) * (1 + (Overlap_Percentage / 100))`   



Some readings which contributed to this estimation:  
1. [Wildland fire growth prediction method based on Multiple Overlapping Solution](https://www.sciencedirect.com/science/article/abs/pii/S1877750310000463)
2. [Vegetation response to a short interval between high-severity wildfires in a mixed-evergreen forest](https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/j.1365-2745.2008.01456.x)
3. [Integrating multiple factors to optimize watchtower deployment for wildfire detection](https://www.sciencedirect.com/science/article/pii/S0048969720330783?casa_token=RwM-EUbbC6MAAAAA:XmccsTeNzvDZeBSFQOWeFB2Ew78fxUXJemoPj1xep32vsfMeOrAExhkgiTSBdra3zri671-yGF4)

4. [Fire behavior and smoke modeling: Model improvement and measurement needs for next-generation smoke research and forecasting systems](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7336523/#:~:text=Fire%20behavior%20and%20smoke%20models%20are%20numerical%20tools%20that%20provide,of%20wildland%20fires%20and%20develop)

# Section 2: Function to calculate Overlap Factor 

Extracting information about the number of years it has been since previous burn and the overlap percentage present in the OverlapFlags column

def calculate_overlap_factor(overlap_flags):
    if pd.isna(overlap_flags):
        return 0  

    if "No overlap information available" in overlap_flags:
        return 0 
    year_pattern = r'Wildfire in (\d{4}) overlaps with a Wildfire that occurred in (\d{4})'
    overlap_pattern = r'overlaps by ([\d.]+)%'

    years_match = re.search(year_pattern, overlap_flags)
    overlap_match = re.search(overlap_pattern, overlap_flags)

    if years_match and overlap_match:
        current_year = int(years_match.group(1))
        previous_year = int(years_match.group(2))
        overlap_percentage = float(overlap_match.group(1))

        years_since_previous_burn = current_year - previous_year

        if years_since_previous_burn < 0:
            years_since_previous_burn = 0  

        overlap_factor = (years_since_previous_burn + 1) * (1 + (overlap_percentage / 100))
        return years_since_previous_burn, overlap_percentage, overlap_factor
    else:
        return 0, 0, 0  

Computing Overlap Factor

filter_df['Years_Since_Previous_Burn'] = filter_df['OverlapFlags'].apply(calculate_overlap_factor)
filter_df['Overlap_Percentage'] = filter_df['OverlapFlags'].apply(calculate_overlap_factor)
filter_df['Overlap_Factor'] = filter_df['OverlapFlags'].apply(calculate_overlap_factor)

filter_df.head()

filter_df.count()

Sanity checking the Overlap Factor

filter_df['Years_Since_Previous_Burn'] = filter_df['Years_Since_Previous_Burn'].apply(lambda x: x[0] if isinstance(x, tuple) else x)
filter_df['Overlap_Percentage'] = filter_df['Overlap_Percentage'].apply(lambda x: x[0] if isinstance(x, tuple) else x)
filter_df['Overlap_Factor'] = filter_df['Overlap_Factor'].apply(lambda x: x[0] if isinstance(x, tuple) else x)

print("Number of rows with Overlap_Factor equal to 0:", len(filter_df[filter_df['Overlap_Factor'] == 0]))
print("Number of rows with negative Overlap_Factor:", len(filter_df[filter_df['Overlap_Factor'] < 0]))
print("Number of rows with positive Overlap_Factor:", len(filter_df[filter_df['Overlap_Factor'] > 0]))

# Section 3: Calculating the Smoke Factor

Checking the data in the GISAcres and Shortest_Distance columns

gisacres_range = (filter_df['GISAcres'].min(), filter_df['GISAcres'].max())
shortest_distance_range = (filter_df['shortest_dist'].min(), filter_df['shortest_dist'].max())

print("Range of GISAcres: ({}, {})".format(gisacres_range[0], gisacres_range[1]))
print("Range of Shortest_Distance: ({}, {})".format(shortest_distance_range[0], shortest_distance_range[1]))

Coming up with acre classification and distance classification 

acre_boundaries = [0, 50000, 500000, filter_df['GISAcres'].max()]
dist_boundaries = [0, 200, 600, filter_df['shortest_dist'].max()]

acre_labels = ['small fire', 'medium fire', 'large fire']
dist_labels = ['close fire', 'intermediate fire', 'far fire']

filter_df['acre classification'] = pd.cut(filter_df['GISAcres'], bins=acre_boundaries, labels=acre_labels)
filter_df['dist classification'] = pd.cut(filter_df['shortest_dist'], bins=dist_boundaries, labels=dist_labels)

filter_df.head()

Assigning smoke factor values to combinations of acre and distance classification

smoke_factor_mapping = {
    ('small fire', 'close fire'): 0.1,             #smaller fires close to the city
    ('small fire', 'intermediate fire'): 0.2,      #Smaller fires at an intermediate distance
    ('small fire', 'far fire'): 0.01,              #Smaller fires far from the city (minimum value)
    ('medium fire', 'close fire'): 0.5,            #Medium-sized fires close to the city
    ('medium fire', 'intermediate fire'): 0.4,     #Medium-sized fires at an intermediate distance
    ('medium fire', 'far fire'): 0.3,              #Medium-sized fires far from the city
    ('large fire', 'close fire'): 0.9,             #Larger fires close to the city (maximum value)
    ('large fire', 'intermediate fire'): 0.8,      #Larger fires at an intermediate distance
    ('large fire', 'far fire'): 0.7,               #Larger fires far from the city
}

filter_df['Smoke_Factor'] = filter_df.apply(lambda row: smoke_factor_mapping.get((row['acre classification'], row['dist classification'])), axis=1)

filter_df.head()

Applying final formula to calculate the Smoke Estimate

# Section 4: Calculating the Smoke Estimate

filter_df['Smoke_Estimate'] = (filter_df['GISAcres'] / filter_df['shortest_dist']) * filter_df['Smoke_Factor'] * (1 + filter_df['Overlap_Factor'])
filter_df

selected_columns = ['FireYear', 'GISAcres', 'shortest_dist', 'Smoke_Estimate']
filter_df = filter_df[selected_columns]

filter_df.to_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/intermediate_data/smoke_est_without_agg.csv', index=False)

filter_df

Grouping the mean smoke estimate by year and taking the mean of the remaining columns

# Section 5: Final Dataframe with Smoke Estimate Data grouped by year

selected_columns = ['FireYear', 'GISAcres', 'shortest_dist', 'Smoke_Estimate']
smoke_est_df = filter_df[selected_columns]
smoke_est_df = smoke_est_df.groupby('FireYear').mean().reset_index()
smoke_est_df

smoke_est_df.to_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/intermediate_data/smoke_est.csv', index=False)
