==> src/5-Data Cleaning.ipynb <==
This notebook contains the data cleaning of the metrics chosen for Part 2 Analysis. These metrics are:
1. Income and Industry data for Salina
2. Unemployment and employment data for Salina

# Section 1: Importing necessary modules

import pandas as pd
import numpy as np
import glob
from scipy.optimize import nnls
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score

# Section 2: Cleaning Income and Industry data for Salina 

The data resides in twelve csv files named 2010_Census.csv, 2011_Census.csv,,..., 2021_Census.csv. We iterate through each of these and perform appropriate data cleaning.  

This step also involves removing all columns with '(X)' values or all null values, and those starting with 'Percent' or 'Unnamed'. Then, we select the columns related to Income and Industry by searching for these strings within the columns and finally, transposing the dataframes to then add Metrics and Sub Metric columns. These two new columns are then populated with the appropriate values (Metric - either Income or Industry, Submetric - different classes of Income/Industry).   

Finally, we rename and drop unnecessary columns and then set the corresponfing year.

def import_clean_data(year):
    file_path = f'/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/data_sources/{year}_Census.csv'
    data = pd.read_csv(file_path, header=1)
    
    # Data cleaning steps
    data = data.drop(['Geography', 'Geographic Area Name'], axis=1)

    # Remove columns with all '(X)' values or all null values
    columns_to_drop_1 = data.columns[data.eq('(X)').all()]
    columns_to_drop_2 = data.columns[data.isnull().all()]
    data.drop(columns=columns_to_drop_1, inplace=True)
    data.drop(columns=columns_to_drop_2, inplace=True)

    # Drop columns starting with 'Percent' or 'Unnamed:'
    columns_to_drop_3 = [col for col in data.columns if col.startswith(('Percent', 'Unnamed:'))]
    data.drop(columns=columns_to_drop_3, inplace=True)

    # Select columns related to 'INCOME' or 'INDUSTRY'
    columns_to_keep = [col for col in data.columns if 'INCOME' in col or 'INDUSTRY' in col]
    data = data[columns_to_keep]

    # Transpose the DataFrame, add 'Metrics', 'Sub Metric' columns
    data = data.T
    data['Metrics'] = np.nan
    data['Sub Metric'] = np.nan
    data = data.reset_index()

    # Classify 'Metrics' and 'Sub Metric' based on column names
    for col in data['index']:
        if 'INDUSTRY' in col:
            data.loc[data['index'] == col, 'Metrics'] = 'INDUSTRY'
            data.loc[data['index'] == col, 'Sub Metric'] = col.rsplit("!!", 1)[-1]
        elif 'INCOME' in col:
            data.loc[data['index'] == col, 'Metrics'] = 'INCOME'
            data.loc[data['index'] == col, 'Sub Metric'] = col.rsplit("!!", 1)[-1]

    # Rename columns, drop unnecessary columns, and set corr year
    data = data.rename(columns={0: 'Salina_Count'})
    data = data.drop('index', axis=1)
    data['Year'] = year
    data.drop_duplicates(subset='Sub Metric', keep='first', inplace=True)

    return data

We create df_income_industry, which is now the final dataframe containing values for income and employment industry in the city of salina from years 2011 to 2021.

# Create an empty DataFrame to store concatenated data
df_income_industry = pd.DataFrame()

# Loop through years from 2011 to 2021 and process data
for year in range(2011, 2022):
    data_year = import_clean_data(year)
    data_name = f'data_{year}'
    globals()[data_name] = data_year  # Create individual DataFrames for each year
    df_income_industry = pd.concat([df_income_industry, data_year])  # Concatenate data

# Reset index for the combined DataFrame
df_income_industry = df_income_industry.reset_index(drop=True)

df_income_industry

df_income_industry.to_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/intermediate_data/income_industry.csv')

# Checking to see if there are any null values
null_count = df_income_industry.isnull().sum()
non_null_count = df_income_industry.notnull().sum()
total_count = df_income_industry.count()

count_summary = pd.DataFrame({
    'Null Values': null_count,
    'Non-Null Values': non_null_count,
    'Total Values': total_count
})
print(count_summary)

# Section 3: Cleaning Employment and Unemployment data for Salina

We firstly import the employment data for Salina and remove the unnecessary columns and only retain year and count of employment in the city.

# Importing the Employment data
df_employment = pd.read_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/data_sources/Salina_Monthly_Employment.csv')
df_employment

df_employment = df_employment.drop(['Series ID', 'Label','Period'], axis=1)
df_employment = df_employment.groupby('Year').agg({'Value': 'mean'}).reset_index()
df_employment = df_employment.rename(columns={'Value': 'Salina_Employment'})

df_employment.head()

Similarly, we import the unemployment data for Salina and remove the unnecessary columns and only retain year and count of unemployment in the city.

# Importing the Unemployment data
df_unemployment = pd.read_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/data_sources/Salina_Monthly_Unemployment.csv')

df_unemployment = df_unemployment.drop(['Series ID', 'Label','Period'], axis=1)
df_unemployment = df_unemployment.groupby('Year').agg({'Value': 'mean'}).reset_index()
df_unemployment = df_unemployment.rename(columns={'Value': 'Salina_Unemployment'})

df_unemployment.head()

# Section 4: Merging both the employment and unemployment data

df_employment_unemployment = pd.merge(df_employment, df_unemployment, on='Year', how='inner')
df_employment_unemployment.head()

df_employment_unemployment contains the final data for unemployment and unemployment for Salina

# Section 5: Final Analysis Datasets

df_smoke=pd.read_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/intermediate_data/pred_df.csv')

df_smoke_pred_1 = pd.merge(df_employment_unemployment, df_smoke, on='Year', how='inner')
df_smoke_pred_1 = df_smoke_pred_1.drop(['shortest_dist','GISAcres'], axis=1)
df_smoke_pred_1.head()

df_smoke_pred_1.to_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/intermediate_data/analysis_1.csv', index=False)

df_smoke_pred_2 = pd.merge(df_income_industry, df_smoke, on='Year', how='inner')
df_smoke_pred_2 = df_smoke_pred_2.drop(['shortest_dist','GISAcres'], axis=1)
df_smoke_pred_2.head()

df_smoke_pred_2.to_csv('/Users/aviva/Desktop/HCDS Project/salina_wildfire_analysis/intermediate_data/analysis_2.csv', index=False)
